Create loop.txt
Save this as a .exe This program uses a TWebBrowser component to embed a video player and loops a YouTube video by setting its loop attribute to true using JavaScript. To stop the video player, you can click a button that sets the TWebBrowser's URL to about:blank. Note that this program may not work as expected with all embedded video players, as different players may have different APIs and behaviors.